Optimize MakePathList and GetShellEscapedString
MakePathList accounts for ~17% of the runtime of
manifest_parser_perftest.

Iterating the list of paths is cheap, and repeated string appends
can be expensive if they have to reallocate.  Iterate twice, once
to determine the total size to reserve in the output string, and
once to do the string appends.

Replace the IsKnownShellSafeCharacter switch statement with a lookup
table that is statically initialized.

Reserve the expected size of the escaped string, assuming it has no
extra single quotes in it.

Use std::find to find single quotes instead of iterating manually.

The first two reduce the number of cycles used by
manifest_parser_perftest by 5%, as measured by
perf stat -r 5 -e cycles:u,instructions:u ./manifest_parser_perftest

The second two are not triggered by manifest_parser_perftest, and will
not be triggered by most manifests.